==> 2-descriptive-outcomes/stunting/0-run-descriptive-outcomes-stunting.R <==

rm(list=ls())
source(paste0(here::here(), "/0-config.R"))

source(here("2-descriptive-outcomes/stunting/0-make-who-growth-vel.R"))
source(here("2-descriptive-outcomes/stunting/1_growth_velocity.R"))
source(here("2-descriptive-outcomes/stunting/2_stunt_calc_outcomes.R"))
source(here("2-descriptive-outcomes/stunting/3_pool_velocity_outcomes.R"))
