==> 2-descriptive-outcomes/stunting/3_pool_velocity_outcomes.R <==
################################################################################
# ki longitudinal manuscripts
# stunting analysis

# pool velocity outcomes
################################################################################


# load packages
rm(list=ls())
source(paste0(here::here(), "/0-config.R"))


source(paste0(here::here(),"/0-project-functions/0_descriptive_epi_shared_functions.R"))
source(paste0(here::here(),"/0-project-functions/0_descriptive_epi_stunt_functions.R"))

d <- readRDS(paste0(ghapdata_dir, "velocity_longfmt_rf.rds"))

#-------------------------------------------------------------------------------

#Summarize N's in study
d %>% group_by(studyid, country, subjid) %>% slice(1) %>% 
  ungroup() %>% summarize(N=n())
d %>% group_by(studyid, country, subjid) %>% slice(1) %>%  
  group_by(studyid, country)  %>% summarize(N=n()) %>% as.data.frame()

#Format results for pooling
table(d$diffcat)

d <- d %>% rename(agecat = diffcat) %>%
  mutate(country_cohort=paste0(studyid," ", country)) %>%
  group_by(country_cohort, studyid, country, agecat, ycat, sex)%>%
  summarise(mean=mean(y_rate, na.rm=T), var=var(y_rate, na.rm=T),
            sd=sd(y_rate, na.rm=T), n=n()) %>%
  mutate(ci.lb=mean - 1.96 * sd/sqrt(n), ci.ub=mean + 1.96 * sd/sqrt(n)) 

#-------------------------------------------------------------------------------
# age specific pooled results
#-------------------------------------------------------------------------------
RE_pool <- function(df, ycategory, gender, method = "REML"){
  
  df <- df %>% filter(ycat==ycategory)
  df <- df %>% filter(sex==gender)
  
  agecat = list("0-3 months", "3-6 months",  "6-9 months","9-12 months",
                "12-15 months","15-18 months")
  
  pooled.vel=lapply(agecat,function(x) 
    fit.rma(data=df, yi="mean", vi="var", ni="n", nlab="children",age=x,
            measure = "MN", method=method))

  
  pooled.vel=as.data.frame(do.call(rbind, pooled.vel))

  # age specific pooled results
  age.vel=lapply(agecat,function(x) fit.rma(data=df, 
        yi="mean", vi="var", ni="n", nlab="children",age=x, measure = "MN", method = method))
  
  
  #Bind together cohort specific estimates
  
  cohort.df <- df %>% subset(., select = c(country_cohort, agecat, n, mean, ci.lb, ci.ub)) %>%
    rename(N=n, Mean=mean, Lower.95.CI=ci.lb, Upper.95.CI=ci.ub,
           strata=agecat) %>%
    mutate(pooled=0, nstudies=1)
  
  plotdf <- bind_rows(cohort.df)
  
  
  #Format variables for plotting
  
  #remove grant identifier
  plotdf$country_cohort<- gsub("^k.*?-" , "",plotdf$country_cohort)
  
  simpleCap <- function(x) {
    s <- strsplit(x, " ")[[1]]
    paste(toupper(substring(s, 1,1)), tolower(substring(s, 2)),
          sep="", collapse=" ")
  }
  
  for(i in 1:length(plotdf$country_cohort)){
    plotdf$country_cohort[i] <- simpleCap(plotdf$country_cohort[i])
  }
  
  plotdf$statistic="Velocity"
  plotdf$country_cohort <- factor(plotdf$country_cohort, levels=unique(plotdf$country_cohort))
  plotdf$strata <- factor(plotdf$strata, levels=unique(plotdf$strata))
  
  plotdf$stratacol <- "strata"
 # plotdf$stratacol[plotdf$strata=="Overall"] <- "overall"
 # plotdf$stratacol[plotdf$pooled==1] <- "pooled"
 # plotdf$stratacol[plotdf$strata=="Overall" & plotdf$pooled==1] <- "pooled_unstrat"
  
  plotdf$sex <- gender
  plotdf$ycat <- ycategory
  
  return(plotdf)
}

#-------------------------------------------------------------------------------
# pool results 
#-------------------------------------------------------------------------------

poolhaz_boys <- RE_pool(d, ycategory="haz", gender="Male")
poolhaz_girls <- RE_pool(d, ycategory="haz", gender="Female")
poollencm_boys <- RE_pool(d, ycategory="lencm", gender="Male")
poollencm_girls <- RE_pool(d, ycategory="lencm", gender="Female")

pooled_vel <- rbind(
  poolhaz_boys, poolhaz_girls, poollencm_boys, poollencm_girls
)

saveRDS(pooled_vel, 
        file=paste0(res_dir,"stunting/pool_vel.RDS"))

#-------------------------------------------------------------------------------
# pool results  -- fixed effects
#-------------------------------------------------------------------------------

poolhaz_boys_fe <- RE_pool(d, ycategory="haz", gender="Male", method="FE")
poolhaz_girls_fe <- RE_pool(d, ycategory="haz", gender="Female", method="FE")
poollencm_boys_fe <- RE_pool(d, ycategory="lencm", gender="Male", method="FE")
poollencm_girls_fe <- RE_pool(d, ycategory="lencm", gender="Female", method="FE")

pooled_vel_fe <- rbind(
  poolhaz_boys_fe, poolhaz_girls_fe, poollencm_boys_fe, poollencm_girls_fe
)

saveRDS(pooled_vel_fe, 
        file=paste0(res_dir,"stunting/pool_vel_fe.RDS"))
